Split non-cipher code to utils.go out of ciphers.go
